Added jQuery AJAX to index.html to make backend calls
