--- Controllers/MapsController.cs
@@ -1,66 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using Locate_closest_business.Models;
-using System.Net.Http;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
-using Newtonsoft.Json.Linq;
-
-
-namespace Locate_closest_business.Controllers
-{
-    public partial class MapsController : Controller
-    {
-        public JsonResult MapsNearbySearch([FromQuery]string lat, [FromQuery]string lng, [FromQuery]string category, int searchRadius = 1500, bool opennow = false)
-        {
-            var errorJSON = new
-            {
-                errorMessage = "External_API_Unreachable"
-            };
-
-            try
-            {
-                Task<MapsResponseWrapperModel> task = Task.Run<MapsResponseWrapperModel>(async () => await PerformNearbySearch(lat,lng,category,searchRadius,opennow));
-                task.Wait();
-                MapsResponseWrapperModel APIResponse = task.Result;
-
-                if (APIResponse == null)
-                {
-                    return Json(errorJSON);
-                }
-                return Json(APIResponse);
-            }
-            catch (Exception exp)
-            {
-                if (exp.Message.Contains("No such host is known."))
-                {
-                    Console.WriteLine(exp);
-                    return Json(errorJSON);
-                }
-                else if (exp.Message.Contains("Response status code does not indicate success: 404 (Not Found)"))
-                {
-                    Console.WriteLine(exp);
-                    var incorrectParamsContent = new
-                    {
-                        errorMessage = "Specified_Country_Slug_Not_Found"
-                    };
-                    return Json(incorrectParamsContent);
-                }
-                else
-                {
-                    Console.WriteLine(exp);
-                    var unexpectedErrorContent = new
-                    {
-                        errorMessage = exp.Message
-                    };
-                    return Json(unexpectedErrorContent);
-                }
-            }
-        }
-    }
-}
