+++ Views/Home/Index.cshtml
@@ -7,18 +7,20 @@
 <body>
     <meta name="viewport" content="width=device-width" />
 
-    <script type="text/javascript" src="https://maps.googleapis.com/maps/api/js?key=&libraries=places"></script>
+    <script type="text/javascript" src=@ViewBag.MapsAPISource></script>
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
     <script type="text/javascript">
         var map;
-        var service;
         var infowindow;
         var usersPosition;
+        var searchMarkers = [];
+
         var myStyles = [
             {
                 featureType: "poi",
                 elementType: "labels",
                 stylers: [
-                    { visibility: "off" }
+                    { visibility: "on" }
                 ]
             }
         ];
@@ -42,45 +44,68 @@
         }
 
         function searchStuff() {
-            var request = {
-                location: usersPosition,
-                radius: '2000',
-                type: ['clothing_store'],
-                //openNow: 'true',
-                fields: 'business_status'
-            };
-            map.setZoom(14);
-            service = new google.maps.places.PlacesService(map);
-            service.nearbySearch(request, callback);
+            clearMarkers();
 
+            var pServiceCategory = document.getElementById("essentialServiceOptions").options[document.getElementById("essentialServiceOptions").selectedIndex].text;
+            //alert(usersPosition.lat() + " :: " + usersPosition.lng());
+
+            $.ajax({
+                type: 'GET',
+                url: 'MapsNearbySearch',
+                data: {
+                    "lat": usersPosition.lat(),
+                    "lng" : usersPosition.lng(),
+                    "category" : pServiceCategory,
+                    "searchRadius" : 2500
+                },
+                success: function (returnData)
+                {
+
+                    if (returnData.status === "OK") {
+                        clearMarkers(map);
+                        for (var i = 0; i < returnData.results.length; i++) {
+                            createMarker(returnData.results[i]);
+                        }
+                    }
+                }
+            });
+            
+            map.setZoom(15);
         }
 
-        function callback(results, status) {
+        /*function callback(results, status) {
             if (status == google.maps.places.PlacesServiceStatus.OK) {
                 for (var i = 0; i < results.length; i++) {
                     createMarker(results[i]);
                 }
             }
-        }
+        }*/
 
         function createMarker(place) {
             if (place.business_status == "OPERATIONAL") {
 
-                    var marker = new google.maps.Marker({
-                        map: map,
-                        position: place.geometry.location,
-                        label: place.name
-                    });
+                var marker = new google.maps.Marker({
+                    map: map,
+                    position: place.geometry.location,
+                    label: place.name
+                });
 
-                    google.maps.event.addListener(marker, 'click', function () {
-                        infowindow.setContent(place.business_status);
-                        infowindow.open(map, this);
-                    });
+                google.maps.event.addListener(marker, 'click', function () {
+                    infowindow.setContent(place.business_status);
+                    infowindow.open(map, this);
+                });
+
+                searchMarkers.push(marker);
             }         
         }
+
+        function clearMarkers(map) {
+            for (var i = 0; i < searchMarkers.length; i++) {
+                searchMarkers[i].setMap(null);
+            }
+            searchMarkers = [];
+        }
     </script>
-
-
     
     <div class="card-body">
         <div id="dvMap" class="card" style="width: 100%; height: 800px">
